pyproject.toml
[project]
name = "ezmd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=0.4.0",
    "markitdown>=0.0.1a3",
    "textual>=1.0.0",
]
